+++ packages/e2e-tests/src/utils.ts
@@ -5,6 +5,7 @@ import getPort from 'get-port'
 import path from 'path'
 import fs from 'fs'
 import { DESKTOP_DATA_DIR, getAppDataPath } from '@quiet/common'
+import { config } from 'dotenv'
 
 export const BACKWARD_COMPATIBILITY_BASE_VERSION = '2.0.1' // Pre-latest production version
 const appImagesPath = `${__dirname}/../Quiet`
@@ -45,11 +46,17 @@ export class BuildSetup {
     this.debugPort = await getPort()
   }
 
+  static getEnvFileName() {
+    const { parsed, error } = config()
+    console.log('Dotenv config', { parsed, error })
+    return process.env.FILE_NAME
+  }
+
   private getBinaryLocation() {
     console.log('filename', this.fileName)
     switch (process.platform) {
       case 'linux':
-        return `${__dirname}/../Quiet/${this.fileName ? this.fileName : process.env.FILE_NAME}`
+        return `${__dirname}/../Quiet/${this.fileName ? this.fileName : BuildSetup.getEnvFileName()}`
       case 'win32':
         return `${process.env.LOCALAPPDATA}\\Programs\\@quietdesktop\\Quiet.exe`
       case 'darwin':
@@ -71,7 +78,7 @@ export class BuildSetup {
   }
 
   public getVersionFromEnv() {
-    const envFileName = process.env.FILE_NAME
+    const envFileName = BuildSetup.getEnvFileName()
     if (!envFileName) {
       throw new Error('file name not specified')
     }