+++ packages/types/src/network.ts
@@ -8,8 +8,24 @@ export type InvitationPair = {
   onionAddress: string
 }
 
-export type InvitationData = {
+export enum InvitationDataVersion {
+  v1 = 'v1',
+  v2 = 'v2',
+}
+
+export type InvitationDataV1 = {
+  version?: InvitationDataVersion.v1
   pairs: InvitationPair[]
   psk: string
   ownerOrbitDbIdentity: string
 }
+
+export type InvitationDataV2 = {
+  version?: InvitationDataVersion.v2
+  cid: string
+  token: string
+  serverAddress: string
+  inviterAddress: string
+}
+
+export type InvitationData = InvitationDataV1 | InvitationDataV2